requirements.txt
Django==4.2.3
gunicorn
requests
psycopg2-binary
psycopg2
django-storages
djangorestframework==3.14.0
djangorestframework-simplejwt
django-cors-headers==4.2.0
django-environ==0.10.0
django-user-agents==0.4.0
user-agents==2.2.0